Update 0.0.11
Now text canvas is a class.
This increases the number of code lines.
But it makes it possible not to load the sprite for a new one every time, separating the areas of responsibility of the classes.
In addition, this allows you to modify the class in the future, if it becomes necessary to increase its area of responsibility.

Fixed a bug with finding coordinates for sprats of heroes when resizing the window.

Fixed a bug with a deterioration in the quality of the background when resizing the window.