Update 0.0.94: UI strategy pattern
Simplified the code.
Edited code comments.

Refactoring:
1) InputCommandsReactions:
a) Now the game menus are called by the strategy pattern.
b) The class now contains a menus_collection shared by the InterfaceController and itself.
It is used for the strategy pattern.

2) InterfaceController:
Now uses strategy pattern to work with all menus.

3) BaseMenu:
Added elements for the strategy pattern and the state machine.

4) Now all menus use the singleton pattern.
They are part of the strategy pattern, and have an outline of the state machine pattern.

5) Slightly changed the render class.