Update 0.1.6:  Character first plan render bagfix
Character:
Optimized 'set_pose' method.
Add new method 'get_pose'.
Now it will correct the poses of the characters.

Fixed missing part of the character sprite being clipped to the background when the character is in the first plan.
For this, part of the mathematics of finding coordinates is transferred to scaling.

The image attribute is divided into two entities.