Update 0.1.22: SaveLoad save Buttons
SaveKeeper:
Now the load and save collection operates on the same data cells.

The position of the save button is no longer stored inside the game save .
Now it is generated automatically.

Button:
Added a method to generate a save cell.
Simplified the code.

UI_buttons_calculations.py:
Added a function to calculate the size of the save cell button.