Update 0.2.7: UI buttons Refactor blank
Divided all interface buttons from the god object into independent classes that inherit the button base class.

Now all methods for calculating the coordinates and sizes of the button are independent and have the same name in new button`s classes.

Now the objects are duplicates and are not used in the interface.